Web Design Accessibility, IRC, and removing banners
I have moved the «Tips and tricks for Web accessibility» page so it is now located on my own website - I intend to rewrite the parts that yourdevilfriends made as a way to have what is on owly.fans be fully open-source.

I have added new IRC logs, this time for #neozones on ibera.chat (I might make a new page for this new room? Not sure).

Removed the banners from the landing page as they don't really fit in/work well. I also want to remove the Xenia image from the same page sometime soon, this is as, despite having permission from digitalangel.neocities.org to use the art, what he made is not under an open licence and I want my website to be as open as possible.